refactor: extract block-registry crate from rustsim block implementations
Move all block implementations (constant, data_driven, embedded, function,
plot, pubsub, register, registry, serde_block, state_machine, udp) into a
standalone block-registry crate. Add BlockRegistry trait to module-traits
for catalog/create abstraction. Rustsim now re-exports from block-registry
via a thin mod.rs shim. SimModel tests that require WasmSimPeripherals
moved to rustsim integration tests since block-registry cannot depend on
rustsim.
