+++ block-registry/src/pubsub.rs
@@ -1,19 +1,19 @@
 //! Pub/Sub source and sink blocks for cross-graph or external messaging.
 
-use crate::dataflow::block::{Module, PortDef, PortKind, Tick, Value};
+use module_traits::{Module, PortDef, PortKind, Tick, Value};
 use serde::{Deserialize, Serialize};
-use tsify_next::Tsify;
 
 // ---------------------------------------------------------------------------
 // Config
 // ---------------------------------------------------------------------------
 
-#[derive(Debug, Serialize, Deserialize, Tsify)]
-#[tsify(into_wasm_abi, from_wasm_abi)]
+#[derive(Debug, Serialize, Deserialize)]
+#[cfg_attr(feature = "tsify", derive(tsify_next::Tsify))]
+#[cfg_attr(feature = "tsify", tsify(into_wasm_abi, from_wasm_abi))]
 #[serde(default)]
 pub struct PubSubConfig {
     pub topic: String,
-    #[tsify(type = "\"Float\" | \"Bytes\" | \"Text\" | \"Series\" | \"Any\"")]
+    #[cfg_attr(feature = "tsify", tsify(type = "\"Float\" | \"Bytes\" | \"Text\" | \"Series\" | \"Any\""))]
     pub port_kind: PortKind,
 }
 
@@ -174,8 +174,8 @@ impl Tick for PubSubSourceBlock {
 }
 
 #[allow(dead_code)]
-pub(crate) fn register(reg: &mut Vec<super::registry::BlockRegistration>) {
-    reg.push(super::registry::BlockRegistration {
+pub(crate) fn register(reg: &mut Vec<crate::registry::BlockRegistration>) {
+    reg.push(crate::registry::BlockRegistration {
         block_type: "pubsub_sink",
         display_name: "PubSub Sink",
         category: "I/O",
@@ -184,7 +184,7 @@ pub(crate) fn register(reg: &mut Vec<super::registry::BlockRegistration>) {
             Ok(Box::new(PubSubSinkBlock::from_config(cfg)))
         },
     });
-    reg.push(super::registry::BlockRegistration {
+    reg.push(crate::registry::BlockRegistration {
         block_type: "pubsub_source",
         display_name: "PubSub Source",
         category: "I/O",
@@ -200,6 +200,7 @@ pub(crate) fn register(reg: &mut Vec<super::registry::BlockRegistration>) {
 // ===========================================================================
 
 #[cfg(test)]
+#[allow(clippy::expect_used)]
 mod tests {
     use super::*;
