+++ crates/rustsim/Cargo.toml
@@ -13,6 +13,7 @@ default = []
 mlir = ["mlir-codegen"]
 
 [dependencies]
+block-registry = { path = "../../block-registry", features = ["tsify"] }
 module-traits = { path = "../../module-traits", features = ["tsify"] }
 parser = { path = "../../parser" }
 wasm-bindgen = "0.2"
